PostJob .tsx and .css added to components needs display work
