+++ front-end/src/components/PostJob.css
@@ -1,0 +1,24 @@
+label{
+  background: rgb(23,123,123);
+ text-align: right;
+ width:100%;
+
+  float: left;
+
+ 
+}
+
+.flex-container {
+  background: rgb(23,123,123);
+ text-align: right;
+
+ height: 100%;
+ display: flex;
+margin: 0 auto;
+
+}
+ input {
+   margin-left: 10%;
+   float: right;
+  
+ }
