--- front-end/src/pages/PostJob.css
@@ -1,17 +1,0 @@
- .flex-container  label{
-    display: flex;
-    flex-direction: row;
-    justify-content: flex-end;
-    text-align: right;
-    width: 400px;
-    line-height: 26px;
-    margin-bottom: 10%;
-    float: left;
-  }
-  
-  .flex-container  input {
-    height: 20px;
-    flex: 0 0 200px;
-    margin-left: 10%;
-    float: right;
-  }
